requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
mysql-connector-python==9.3.0
python-dotenv==1.1.0
